+++ src/main/java/io/github/ngspace/hudder/v2runtime/values/V2Values.java
@@ -1,6 +1,6 @@
-package io.github.ngspace.hudder.compilers.v2runtime.values;
+package io.github.ngspace.hudder.v2runtime.values;
 
-import io.github.ngspace.hudder.compilers.v2runtime.AV2Compiler;
+import io.github.ngspace.hudder.v2runtime.AV2Compiler;
 
 public class V2Values {private V2Values() {}
 	//Only after writing 80% of the values did I realize having one class that is all values is bad, I tried to
@@ -25,10 +25,8 @@ public class V2Values {private V2Values() {}
 		
 		return new V2Value(valuee, compiler);
 	}
-
-	//TODO Fix setting a value to a string will make the String lowercase, I've got bigger issues rn.
+	
 	private static V2String string(String value, AV2Compiler compiler) {
-		System.out.println("St: "+value);
 		//Maybe String :)
 		if (!value.startsWith("\"")||!value.endsWith("\"")) return null;
 		
@@ -39,10 +37,8 @@ public class V2Values {private V2Values() {}
 		boolean safe = false;
 		for (int i = 0;i<value.length();i++) {
 			c = value.charAt(i);
-			if (c=='\\'&&!safe) {
-				safe = true;
-				continue;
-			} else {
+			if (c=='\\'&&!safe) safe = true;
+			else {
 				if (c=='"'&&!safe) return null; //Not String ;_;
 				safe = false;
 				string.append(c);
